fte: refactor to reuse crd::s3 anc move resolved struct out of the crd module
